Revert "Change enum entry value type from uint64 to uint32"
This reverts commit 36b3c49d2eb07d205021dbe6f9b7141cb72a599c.

+++ mavgen/mavgen.go
@@ -34,7 +34,7 @@ type Enum struct {
 }
 
 type EnumEntry struct {
-	Value       uint32            `xml:"value,attr"`
+	Value       uint64            `xml:"value,attr"`
 	Name        string            `xml:"name,attr"`
 	Description string            `xml:"description"`
 	Params      []*EnumEntryParam `xml:"param"`
@@ -371,7 +371,7 @@ const ({{range .Entries}}
 		e.Name = UpperCamelCase(e.Name)
 		for i, ee := range e.Entries {
 			if ee.Value == 0 {
-				ee.Value = uint32(i)
+				ee.Value = uint64(i)
 			}
 			ee.Description = strings.Replace(ee.Description, "\n", " ", -1)
 		}
